requirements.txt
alembic
blinker
click
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
Flask-Testing
Flask-WTF
python-decouple
email_validator
